@@ -1,10 +1,24 @@
 module eccs
 
-go 1.16
+go 1.17
 
-replace client => ../../client
+replace github.com/cyber-crypt-com/encryptonize-core/client => ../../client
 
 require (
-	client v0.0.0-00010101000000-000000000000
+	github.com/cyber-crypt-com/encryptonize-core/client v0.0.0-00010101000000-000000000000
 	github.com/spf13/cobra v1.2.1
+)
+
+require (
+	github.com/fullstorydev/grpcurl v1.8.5 // indirect
+	github.com/golang/protobuf v1.5.2 // indirect
+	github.com/inconshreveable/mousetrap v1.0.0 // indirect
+	github.com/jhump/protoreflect v1.10.1 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
+	golang.org/x/net v0.0.0-20211123203042-d83791d6bcd9 // indirect
+	golang.org/x/sys v0.0.0-20211124211545-fe61309f8881 // indirect
+	golang.org/x/text v0.3.7 // indirect
+	google.golang.org/genproto v0.0.0-20211129164237-f09f9a12af12 // indirect
+	google.golang.org/grpc v1.42.0 // indirect
+	google.golang.org/protobuf v1.27.1 // indirect
 )
