@@ -1,9 +1,18 @@
 module github.com/cyber-crypt-com/encryptonize-core/client
 
-go 1.16
+go 1.17
 
 require (
 	github.com/fullstorydev/grpcurl v1.8.5
 	github.com/jhump/protoreflect v1.10.1
 	google.golang.org/grpc v1.42.0
+)
+
+require (
+	github.com/golang/protobuf v1.5.2 // indirect
+	golang.org/x/net v0.0.0-20211123203042-d83791d6bcd9 // indirect
+	golang.org/x/sys v0.0.0-20211124211545-fe61309f8881 // indirect
+	golang.org/x/text v0.3.7 // indirect
+	google.golang.org/genproto v0.0.0-20211129164237-f09f9a12af12 // indirect
+	google.golang.org/protobuf v1.27.1 // indirect
 )
