@@ -12,34 +12,31 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//go:build encryption
-// +build encryption
+//go:build !encryption
+// +build !encryption
 
 package grpce2e
 
 import (
-	"bytes"
 	"testing"
+
+	"context"
+
+	coreclient "github.com/cyber-crypt-com/encryptonize-core/client"
 )
 
-func TestEncryptAndDecrypt(t *testing.T) {
-	client, err := NewClient(endpoint, certPath)
+// Test that we can't store an object
+func TestEncDisabled(t *testing.T) {
+	client, err := coreclient.NewClient(context.Background(), endpoint, certPath)
 	failOnError("Could not create client", err, t)
-	defer closeClient(client, t)
+	defer client.Close()
 
-	_, err = client.LoginUser(uid, pwd)
+	err = client.LoginUser(uid, pwd)
 	failOnError("Could not log in user", err, t)
 
 	plaintext := []byte("foo")
 	associatedData := []byte("bar")
 
-	encryptResponse, err := client.Encrypt(plaintext, associatedData)
-	failOnError("Encrypt operation failed", err, t)
-
-	retrieveResponse, err := client.Decrypt(encryptResponse.Ciphertext, associatedData, encryptResponse.ObjectId)
-	failOnError("Decrypt operation failed", err, t)
-
-	if !bytes.Equal(retrieveResponse.Plaintext, plaintext) {
-		t.Fatalf("Expected plaintext %v but got %v", plaintext, retrieveResponse.Plaintext)
-	}
+	_, err = client.Encrypt(plaintext, associatedData)
+	failOnSuccess("Encrypt operation should have failed", err, t)
 }
