@@ -20,18 +20,20 @@ package grpce2e
 import (
 	"testing"
 
-	"github.com/gofrs/uuid"
+	"context"
 
-	"encryption-service/common"
+	coreclient "github.com/cyber-crypt-com/encryptonize-core/client"
 )
+
+var invalidUID = "6025ea7f-2038-41e0-bc17-aeb2557ded90"
 
 // Test that we can share an object by adding a group to the access object
 func TestShareWithGroup(t *testing.T) {
-	client, err := NewClient(endpoint, certPath)
+	client, err := coreclient.NewClient(context.Background(), endpoint, certPath)
 	failOnError("Could not create client", err, t)
-	defer closeClient(client, t)
+	defer client.Close()
 
-	_, err = client.LoginUser(uid, pwd)
+	err = client.LoginUser(uid, pwd)
 	failOnError("Could not log in user", err, t)
 
 	// Store an object
@@ -40,28 +42,28 @@ func TestShareWithGroup(t *testing.T) {
 
 	storeResponse, err := client.Store(plaintext, associatedData)
 	failOnError("Store operation failed", err, t)
-	oid := storeResponse.ObjectId
+	oid := storeResponse.ObjectID
 
 	// Create a group
 	createGroupResponse, err := client.CreateGroup(protoUserScopes)
 	failOnError("Group creation failed", err, t)
-	gid := createGroupResponse.GroupId
+	gid := createGroupResponse.GroupID
 
 	// Create another user, add them to the group
 	createUserResponse, err := client.CreateUser(protoUserScopes)
 	failOnError("User creation failed", err, t)
-	uid2 := createUserResponse.UserId
+	uid2 := createUserResponse.UserID
 	pwd2 := createUserResponse.Password
 
-	_, err = client.AddUserToGroup(uid2, gid)
+	err = client.AddUserToGroup(uid2, gid)
 	failOnError("Adding user to group failed", err, t)
 
 	// Add the group to the access object
-	_, err = client.AddPermission(oid, gid)
+	err = client.AddPermission(oid, gid)
 	failOnError("Add permission request failed", err, t)
 
 	// Try to retrieve the object as the other user
-	_, err = client.LoginUser(uid2, pwd2)
+	err = client.LoginUser(uid2, pwd2)
 	failOnError("Could not log in user", err, t)
 
 	_, err = client.Retrieve(oid)
@@ -71,11 +73,11 @@ func TestShareWithGroup(t *testing.T) {
 // Test that a user with the READ scope in one group cannot access an object if the group that gives
 // them access to the object does not have that scope.
 func TestGroupScopes(t *testing.T) {
-	client, err := NewClient(endpoint, certPath)
+	client, err := coreclient.NewClient(context.Background(), endpoint, certPath)
 	failOnError("Could not create client", err, t)
-	defer closeClient(client, t)
+	defer client.Close()
 
-	_, err = client.LoginUser(uid, pwd)
+	err = client.LoginUser(uid, pwd)
 	failOnError("Could not log in user", err, t)
 
 	// Store an object
@@ -84,29 +86,29 @@ func TestGroupScopes(t *testing.T) {
 
 	storeResponse, err := client.Store(plaintext, associatedData)
 	failOnError("Store operation failed", err, t)
-	oid := storeResponse.ObjectId
+	oid := storeResponse.ObjectID
 
 	// Create a group with restricted scopes
-	createGroupResponse, err := client.CreateGroup([]common.Scope{common.Scope_CREATE})
+	createGroupResponse, err := client.CreateGroup([]coreclient.Scope{coreclient.ScopeCreate})
 	failOnError("Group creation failed", err, t)
-	gid := createGroupResponse.GroupId
+	gid := createGroupResponse.GroupID
 
 	// Create another user with wider scopes, add them to the group
 	createUserResponse, err := client.CreateUser(protoUserScopes)
 	failOnError("User creation failed", err, t)
-	uid2 := createUserResponse.UserId
+	uid2 := createUserResponse.UserID
 	pwd2 := createUserResponse.Password
 
-	_, err = client.AddUserToGroup(uid2, gid)
+	err = client.AddUserToGroup(uid2, gid)
 	failOnError("Adding user to group failed", err, t)
 
 	// Add the group to the access object
-	_, err = client.AddPermission(oid, gid)
+	err = client.AddPermission(oid, gid)
 	failOnError("Add permission request failed", err, t)
 
 	// Try to retrieve the object as the other user. It should fail as the group that gives them
 	// access does not have the read scope.
-	_, err = client.LoginUser(uid2, pwd2)
+	err = client.LoginUser(uid2, pwd2)
 	failOnError("Could not log in user", err, t)
 
 	_, err = client.Retrieve(oid)
@@ -115,58 +117,56 @@ func TestGroupScopes(t *testing.T) {
 
 // Test creation of a group wth invalid scopes
 func TestCreateGroupInvalidScopes(t *testing.T) {
-	client, err := NewClient(endpoint, certPath)
+	client, err := coreclient.NewClient(context.Background(), endpoint, certPath)
 	failOnError("Could not create client", err, t)
-	defer closeClient(client, t)
+	defer client.Close()
 
-	_, err = client.LoginUser(uid, pwd)
+	err = client.LoginUser(uid, pwd)
 	failOnError("Could not log in user", err, t)
 
-	// Create a group with invalid scopes
-	_, err = client.CreateGroup([]common.Scope{42})
+	invalidScope := []coreclient.Scope{42}
+	_, err = client.CreateGroup(invalidScope)
 	failOnSuccess("Expected group creation to fail", err, t)
 }
 
 // Test adding a user to an invalid group
 func TestAddUserToGroupInvalidGroup(t *testing.T) {
-	client, err := NewClient(endpoint, certPath)
+	client, err := coreclient.NewClient(context.Background(), endpoint, certPath)
 	failOnError("Could not create client", err, t)
-	defer closeClient(client, t)
+	defer client.Close()
 
-	_, err = client.LoginUser(uid, pwd)
+	err = client.LoginUser(uid, pwd)
 	failOnError("Could not log in user", err, t)
 
-	// Try to add use to an invalid group
-	_, err = client.AddUserToGroup(uid, uuid.Must(uuid.NewV4()).String())
+	err = client.AddUserToGroup(uid, invalidUID)
 	failOnSuccess("Expected adding user to group to fail", err, t)
 }
 
 // Test adding an invalid user to a group
 func TestAddUserToGroupInvalidUser(t *testing.T) {
-	client, err := NewClient(endpoint, certPath)
+	client, err := coreclient.NewClient(context.Background(), endpoint, certPath)
 	failOnError("Could not create client", err, t)
-	defer closeClient(client, t)
+	defer client.Close()
 
-	_, err = client.LoginUser(uid, pwd)
+	err = client.LoginUser(uid, pwd)
 	failOnError("Could not log in user", err, t)
 
 	// Create a group
 	createGroupResponse, err := client.CreateGroup(protoUserScopes)
 	failOnError("Group creation failed", err, t)
-	gid := createGroupResponse.GroupId
+	gid := createGroupResponse.GroupID
 
-	// Try to add invalid user to the group
-	_, err = client.AddUserToGroup(uuid.Must(uuid.NewV4()).String(), gid)
+	err = client.AddUserToGroup(invalidUID, gid)
 	failOnSuccess("Expected adding user to group to fail", err, t)
 }
 
 // Test that we can remove a user from a group, stopping them from accessing an object
 func TestRemoveUserFromGroup(t *testing.T) {
-	client, err := NewClient(endpoint, certPath)
+	client, err := coreclient.NewClient(context.Background(), endpoint, certPath)
 	failOnError("Could not create client", err, t)
-	defer closeClient(client, t)
+	defer client.Close()
 
-	_, err = client.LoginUser(uid, pwd)
+	err = client.LoginUser(uid, pwd)
 	failOnError("Could not log in user", err, t)
 
 	// Store an object
@@ -175,35 +175,35 @@ func TestRemoveUserFromGroup(t *testing.T) {
 
 	storeResponse, err := client.Store(plaintext, associatedData)
 	failOnError("Store operation failed", err, t)
-	oid := storeResponse.ObjectId
+	oid := storeResponse.ObjectID
 
 	// Create a group
 	createGroupResponse, err := client.CreateGroup(protoUserScopes)
 	failOnError("Group creation failed", err, t)
-	gid := createGroupResponse.GroupId
+	gid := createGroupResponse.GroupID
 
 	// Create another user, add them to the group
 	createUserResponse, err := client.CreateUser(protoUserScopes)
 	failOnError("User creation failed", err, t)
-	uid2 := createUserResponse.UserId
+	uid2 := createUserResponse.UserID
 	pwd2 := createUserResponse.Password
 
-	_, err = client.AddUserToGroup(uid2, gid)
+	err = client.AddUserToGroup(uid2, gid)
 	failOnError("Adding user to group failed", err, t)
 
 	// Add the group to the access object
-	_, err = client.AddPermission(oid, gid)
+	err = client.AddPermission(oid, gid)
 	failOnError("Add permission request failed", err, t)
 
 	// Try to retrieve the object as the other user
-	_, err = client.LoginUser(uid2, pwd2)
+	err = client.LoginUser(uid2, pwd2)
 	failOnError("Could not log in user", err, t)
 
 	_, err = client.Retrieve(oid)
 	failOnError("Retrieving object failed", err, t)
 
 	// Remove the user from the group and check that they no longer have access
-	_, err = client.RemoveUserFromGroup(uid2, gid)
+	err = client.RemoveUserFromGroup(uid2, gid)
 	failOnError("Removing user from group failed", err, t)
 
 	_, err = client.Retrieve(oid)
@@ -212,65 +212,64 @@ func TestRemoveUserFromGroup(t *testing.T) {
 
 // Test removing an invalid user from a group
 func TestRemoveUserFromGroupInvalidUser(t *testing.T) {
-	client, err := NewClient(endpoint, certPath)
+	client, err := coreclient.NewClient(context.Background(), endpoint, certPath)
 	failOnError("Could not create client", err, t)
-	defer closeClient(client, t)
+	defer client.Close()
 
-	_, err = client.LoginUser(uid, pwd)
+	err = client.LoginUser(uid, pwd)
 	failOnError("Could not log in user", err, t)
 
 	// Create a group
 	createGroupResponse, err := client.CreateGroup(protoUserScopes)
 	failOnError("Group creation failed", err, t)
-	gid := createGroupResponse.GroupId
+	gid := createGroupResponse.GroupID
 
-	// Try to remove an invalid user from the group
-	_, err = client.RemoveUserFromGroup(uuid.Must(uuid.NewV4()).String(), gid)
+	err = client.RemoveUserFromGroup(invalidUID, gid)
 	failOnSuccess("Expected removing user from group to fail", err, t)
 }
 
 // Test that removing a user from a group twice does not fail
 func TestRemoveUserFromGroupTwice(t *testing.T) {
-	client, err := NewClient(endpoint, certPath)
+	client, err := coreclient.NewClient(context.Background(), endpoint, certPath)
 	failOnError("Could not create client", err, t)
-	defer closeClient(client, t)
+	defer client.Close()
 
-	_, err = client.LoginUser(uid, pwd)
+	err = client.LoginUser(uid, pwd)
 	failOnError("Could not log in user", err, t)
 
 	// Create a group, add user to it
 	createGroupResponse, err := client.CreateGroup(protoUserScopes)
 	failOnError("Group creation failed", err, t)
-	gid := createGroupResponse.GroupId
+	gid := createGroupResponse.GroupID
 
-	_, err = client.AddUserToGroup(uid, gid)
+	err = client.AddUserToGroup(uid, gid)
 	failOnError("Adding user to group failed", err, t)
 
 	// Try to remove the user from the group twice
-	_, err = client.RemoveUserFromGroup(uid, gid)
+	err = client.RemoveUserFromGroup(uid, gid)
 	failOnError("Removing user from group failed", err, t)
 
-	_, err = client.RemoveUserFromGroup(uid, gid)
+	err = client.RemoveUserFromGroup(uid, gid)
 	failOnError("Removing user from group failed", err, t)
 }
 
 // Test that we can remove a user from their default group
 func TestRemoveUserFromDefaultGroup(t *testing.T) {
-	client, err := NewClient(endpoint, certPath)
+	client, err := coreclient.NewClient(context.Background(), endpoint, certPath)
 	failOnError("Could not create client", err, t)
-	defer closeClient(client, t)
+	defer client.Close()
 
-	_, err = client.LoginUser(uid, pwd)
+	err = client.LoginUser(uid, pwd)
 	failOnError("Could not log in user", err, t)
 
 	// Create a user, a default group is automatically created
 	createUserResponse, err := client.CreateUser(protoUserScopes)
 	failOnError("User creation failed", err, t)
-	uid2 := createUserResponse.UserId
+	uid2 := createUserResponse.UserID
 	pwd2 := createUserResponse.Password
 
 	// Store an object as the new user
-	_, err = client.LoginUser(uid2, pwd2)
+	err = client.LoginUser(uid2, pwd2)
 	failOnError("Could not log in user", err, t)
 
 	plaintext := []byte("foo")
@@ -278,10 +277,10 @@ func TestRemoveUserFromDefaultGroup(t *testing.T) {
 
 	storeResponse, err := client.Store(plaintext, associatedData)
 	failOnError("Store operation failed", err, t)
-	oid := storeResponse.ObjectId
+	oid := storeResponse.ObjectID
 
 	// Remove user from their default group. They are now a member of no group
-	_, err = client.RemoveUserFromGroup(uid2, uid2)
+	err = client.RemoveUserFromGroup(uid2, uid2)
 	failOnError("Removing user from group failed", err, t)
 
 	// They should be unable to access the object
