@@ -18,46 +18,50 @@
 package grpce2e
 
 import (
-	"reflect"
 	"testing"
+
+	"context"
+	"reflect"
+
+	coreclient "github.com/cyber-crypt-com/encryptonize-core/client"
 )
 
 // Test that a user can remove themselves from the object ACL and cannot decrypt the object afterwards
 func TestDecryptSameUserWithoutPermissions(t *testing.T) {
-	client, err := NewClient(endpoint, certPath)
+	client, err := coreclient.NewClient(context.Background(), endpoint, certPath)
 	failOnError("Could not create client", err, t)
-	defer closeClient(client, t)
+	defer client.Close()
 
-	_, err = client.LoginUser(uid, pwd)
+	err = client.LoginUser(uid, pwd)
 	failOnError("Could not create client", err, t)
 
 	encResponse, err := client.Encrypt([]byte("foo"), []byte("bar"))
 	failOnError("Encrypt operation failed", err, t)
-	oid := encResponse.ObjectId
+	oid := encResponse.ObjectID
 
 	// Remove self from object ACL
-	_, err = client.RemovePermission(oid, uid)
+	err = client.RemovePermission(oid, uid)
 	failOnError("Removing permissions from self failed", err, t)
 
 	// Try to fetch object without permissions
-	_, err = client.Decrypt(encResponse.Ciphertext, encResponse.AssociatedData, oid)
+	_, err = client.Decrypt(oid, encResponse.Ciphertext, encResponse.AssociatedData)
 	failOnSuccess("User should not be able to decrypt object with no permissions", err, t)
 }
 
 // Test that an encrypted object can be retrieved by another user with permissions
 func TestShareEncryptedObjectWithUser(t *testing.T) {
 	// Create another user to share the object with
-	client, err := NewClient(endpoint, certPath)
+	client, err := coreclient.NewClient(context.Background(), endpoint, certPath)
 	failOnError("Could not create client", err, t)
-	defer closeClient(client, t)
+	defer client.Close()
 
-	_, err = client.LoginUser(uid, pwd)
+	err = client.LoginUser(uid, pwd)
 	failOnError("Could not log in user", err, t)
 
 	createUserResponse, err := client.CreateUser(protoUserScopes)
 	failOnError("Create user request failed", err, t)
 
-	uid2 := createUserResponse.UserId
+	uid2 := createUserResponse.UserID
 	pwd2 := createUserResponse.Password
 
 	// Encrypt an object
@@ -65,24 +69,24 @@ func TestShareEncryptedObjectWithUser(t *testing.T) {
 	associatedData := []byte("bar")
 	encryptResponse, err := client.Encrypt(plaintext, associatedData)
 	failOnError("Encrypt operation failed", err, t)
-	oid := encryptResponse.ObjectId
+	oid := encryptResponse.ObjectID
 
 	// Add another user to object permission list
-	_, err = client.AddPermission(oid, uid2)
+	err = client.AddPermission(oid, uid2)
 	failOnError("Add permission request failed", err, t)
 
 	// Try to retrieve object with another user
-	_, err = client.LoginUser(uid2, pwd2)
+	err = client.LoginUser(uid2, pwd2)
 	failOnError("Could not log in user", err, t)
 
-	_, err = client.Decrypt(encryptResponse.Ciphertext, encryptResponse.AssociatedData, oid)
+	_, err = client.Decrypt(oid, encryptResponse.Ciphertext, encryptResponse.AssociatedData)
 	failOnError("Authorized user could not decrypt object created by another user", err, t)
 
 	// Try to retrieve object with the original user
-	_, err = client.LoginUser(uid, pwd)
+	err = client.LoginUser(uid, pwd)
 	failOnError("Could not log in user", err, t)
 
-	_, err = client.Decrypt(encryptResponse.Ciphertext, encryptResponse.AssociatedData, oid)
+	_, err = client.Decrypt(oid, encryptResponse.Ciphertext, encryptResponse.AssociatedData)
 	failOnError("Object creator could not decrypt object", err, t)
 }
 
@@ -90,17 +94,17 @@ func TestShareEncryptedObjectWithUser(t *testing.T) {
 // retrieved by another user without permissions
 func TestDecryptWithoutPermissions(t *testing.T) {
 	// Create another user to share the object with
-	client, err := NewClient(endpoint, certPath)
+	client, err := coreclient.NewClient(context.Background(), endpoint, certPath)
 	failOnError("Could not create client", err, t)
-	defer closeClient(client, t)
+	defer client.Close()
 
-	_, err = client.LoginUser(uid, pwd)
+	err = client.LoginUser(uid, pwd)
 	failOnError("Could not log in user", err, t)
 
 	createUserResponse, err := client.CreateUser(protoUserScopes)
 	failOnError("Create user request failed", err, t)
 
-	uid2 := createUserResponse.UserId
+	uid2 := createUserResponse.UserID
 	pwd2 := createUserResponse.Password
 
 	// Encrypt an object
@@ -108,13 +112,13 @@ func TestDecryptWithoutPermissions(t *testing.T) {
 	associatedData := []byte("bar")
 	encryptResponse, err := client.Encrypt(plaintext, associatedData)
 	failOnError("Encrypt operation failed", err, t)
-	oid := encryptResponse.ObjectId
+	oid := encryptResponse.ObjectID
 
 	// Try to retrieve object with another user
-	_, err = client.LoginUser(uid2, pwd2)
+	err = client.LoginUser(uid2, pwd2)
 	failOnError("Could not log in user", err, t)
 
-	_, err = client.Decrypt(encryptResponse.Ciphertext, encryptResponse.AssociatedData, oid)
+	_, err = client.Decrypt(oid, encryptResponse.Ciphertext, encryptResponse.AssociatedData)
 	failOnSuccess("Unauthorized user should not be able to decrypt object", err, t)
 }
 
@@ -122,24 +126,24 @@ func TestDecryptWithoutPermissions(t *testing.T) {
 // If user A grants access to user B, then user B should be able to grant access to user C
 func TestPermissionTransitivityEnc(t *testing.T) {
 	// Create admin client for user creation
-	client, err := NewClient(endpoint, certPath)
+	client, err := coreclient.NewClient(context.Background(), endpoint, certPath)
 	failOnError("Could not create client", err, t)
-	defer closeClient(client, t)
+	defer client.Close()
 
-	_, err = client.LoginUser(uid, pwd)
+	err = client.LoginUser(uid, pwd)
 	failOnError("Could not create client", err, t)
 
 	// Create users B and C
 	createUserResponse, err := client.CreateUser(protoUserScopes)
 	failOnError("Create user request failed", err, t)
 
-	uid2 := createUserResponse.UserId
+	uid2 := createUserResponse.UserID
 	pwd2 := createUserResponse.Password
 
 	createUserResponse, err = client.CreateUser(protoUserScopes)
 	failOnError("Create user request failed", err, t)
 
-	uid3 := createUserResponse.UserId
+	uid3 := createUserResponse.UserID
 	pwd3 := createUserResponse.Password
 
 	// Encrypt an object
@@ -147,24 +151,24 @@ func TestPermissionTransitivityEnc(t *testing.T) {
 	associatedData := []byte("bar")
 	encryptResponse, err := client.Encrypt(plaintext, associatedData)
 	failOnError("Encrypt operation failed", err, t)
-	oid := encryptResponse.ObjectId
+	oid := encryptResponse.ObjectID
 
 	// Grant permissions to user B
-	_, err = client.AddPermission(oid, uid2)
+	err = client.AddPermission(oid, uid2)
 	failOnError("Add permission request failed", err, t)
 
 	// Use user B to grant permissions to user C
-	_, err = client.LoginUser(uid2, pwd2)
+	err = client.LoginUser(uid2, pwd2)
 	failOnError("Could not create client", err, t)
 
-	_, err = client.AddPermission(oid, uid3)
+	err = client.AddPermission(oid, uid3)
 	failOnError("Add permission request failed", err, t)
 
 	// Check that user C has access to object
-	_, err = client.LoginUser(uid3, pwd3)
+	err = client.LoginUser(uid3, pwd3)
 	failOnError("Could not create client", err, t)
 
-	_, err = client.Decrypt(encryptResponse.Ciphertext, encryptResponse.AssociatedData, oid)
+	_, err = client.Decrypt(oid, encryptResponse.Ciphertext, encryptResponse.AssociatedData)
 	failOnError("Could not decrypt object", err, t)
 }
 
@@ -172,18 +176,18 @@ func TestPermissionTransitivityEnc(t *testing.T) {
 // and that add/remove permission inflicts the outcome of get permissions
 func TestGetPermissionsEnc(t *testing.T) {
 	// Create admin client for user creation
-	client, err := NewClient(endpoint, certPath)
+	client, err := coreclient.NewClient(context.Background(), endpoint, certPath)
 	failOnError("Could not create client", err, t)
-	defer closeClient(client, t)
+	defer client.Close()
 
-	_, err = client.LoginUser(uid, pwd)
+	err = client.LoginUser(uid, pwd)
 	failOnError("Could not log in user", err, t)
 
 	// Create user 2
 	createUserResponse, err := client.CreateUser(protoUserScopes)
 	failOnError("Create user request failed", err, t)
 
-	uid2 := createUserResponse.UserId
+	uid2 := createUserResponse.UserID
 	pwd2 := createUserResponse.Password
 
 	// Encrypt an object
@@ -191,69 +195,69 @@ func TestGetPermissionsEnc(t *testing.T) {
 	associatedData := []byte("bar")
 	encryptResponse, err := client.Encrypt(plaintext, associatedData)
 	failOnError("Encrypt operation failed", err, t)
-	oid := encryptResponse.ObjectId
+	oid := encryptResponse.ObjectID
 
 	//Check that user 2 cannot get permissions from object, without permissions
-	_, err = client.LoginUser(uid2, pwd2)
+	err = client.LoginUser(uid2, pwd2)
 	failOnError("Could not create client", err, t)
 
 	_, err = client.GetPermissions(oid)
 	failOnSuccess("Unauthorized user should not be able to access object permissions", err, t)
 
 	// Grant permissions to user 2
-	_, err = client.LoginUser(uid, pwd)
+	err = client.LoginUser(uid, pwd)
 	failOnError("Could not create client", err, t)
 
-	_, err = client.AddPermission(oid, uid2)
+	err = client.AddPermission(oid, uid2)
 	failOnError("Add permission request failed", err, t)
 
 	getPermissionsResponse1, err := client.GetPermissions(oid)
 	failOnError("Could not get permissions", err, t)
 
-	_, err = client.LoginUser(uid2, pwd2)
+	err = client.LoginUser(uid2, pwd2)
 	failOnError("Could not create client", err, t)
 
 	getPermissionsResponse2, err := client.GetPermissions(oid)
 	failOnError("Could not get permissions", err, t)
 
 	// Check that permissions response contains the right gids
-	ok := find(getPermissionsResponse1.GroupIds, uid)
+	ok := find(getPermissionsResponse1.GroupIDs, uid)
 	if !ok {
-		t.Fatalf("Couldn't find %v in %v", uid, getPermissionsResponse1.GroupIds)
+		t.Fatalf("Couldn't find %v in %v", uid, getPermissionsResponse1.GroupIDs)
 	}
-	ok = find(getPermissionsResponse1.GroupIds, uid2)
+	ok = find(getPermissionsResponse1.GroupIDs, uid2)
 	if !ok {
-		t.Fatalf("Couldn't find %v in %v", uid, getPermissionsResponse1.GroupIds)
+		t.Fatalf("Couldn't find %v in %v", uid, getPermissionsResponse1.GroupIDs)
 	}
 
-	if !reflect.DeepEqual(getPermissionsResponse1.GroupIds, getPermissionsResponse2.GroupIds) {
-		t.Fatalf("Permissions aren't the same: %v vs %v", getPermissionsResponse1.GroupIds, getPermissionsResponse2.GroupIds)
+	if !reflect.DeepEqual(getPermissionsResponse1.GroupIDs, getPermissionsResponse2.GroupIDs) {
+		t.Fatalf("Permissions aren't the same: %v vs %v", getPermissionsResponse1.GroupIDs, getPermissionsResponse2.GroupIDs)
 	}
 
 	// Remove user 2
-	_, err = client.LoginUser(uid, pwd)
+	err = client.LoginUser(uid, pwd)
 	failOnError("Could not create client", err, t)
 
-	_, err = client.RemovePermission(oid, uid2)
+	err = client.RemovePermission(oid, uid2)
 	failOnError("Could not remove permissions", err, t)
 
 	// Check that permissions have been removed
 	getPermissionsResponse1, err = client.GetPermissions(oid)
 	failOnError("Could not get permissions", err, t)
-	ok = find(getPermissionsResponse1.GroupIds, uid)
+	ok = find(getPermissionsResponse1.GroupIDs, uid)
 	if !ok {
-		t.Fatalf("Couldn't find %v in %v", uid, getPermissionsResponse1.GroupIds)
+		t.Fatalf("Couldn't find %v in %v", uid, getPermissionsResponse1.GroupIDs)
 	}
-	ok = find(getPermissionsResponse1.GroupIds, uid2)
+	ok = find(getPermissionsResponse1.GroupIDs, uid2)
 	if ok {
-		t.Fatalf("Found %v in %v", uid, getPermissionsResponse1.GroupIds)
+		t.Fatalf("Found %v in %v", uid, getPermissionsResponse1.GroupIDs)
 	}
 
 	// Check that user 2 doesn't have permissions
-	_, err = client.LoginUser(uid2, pwd2)
+	err = client.LoginUser(uid2, pwd2)
 	failOnError("Could not create client", err, t)
 
-	_, err = client.Decrypt(encryptResponse.Ciphertext, encryptResponse.AssociatedData, oid)
+	_, err = client.Decrypt(oid, encryptResponse.Ciphertext, encryptResponse.AssociatedData)
 	failOnSuccess("Unauthorized user should not be able to decrypt object", err, t)
 }
 
@@ -261,18 +265,18 @@ func TestGetPermissionsEnc(t *testing.T) {
 func TestAddPermissionNoTargetUserEnc(t *testing.T) {
 	nonExistingUser := "00000000-0000-0000-0000-000000000000"
 
-	client, err := NewClient(endpoint, certPath)
+	client, err := coreclient.NewClient(context.Background(), endpoint, certPath)
 	failOnError("Could not create client", err, t)
-	defer closeClient(client, t)
+	defer client.Close()
 
-	_, err = client.LoginUser(uid, pwd)
+	err = client.LoginUser(uid, pwd)
 	failOnError("Could not create client", err, t)
 
 	encryptResponse, err := client.Encrypt([]byte("foo"), []byte("bar"))
 	failOnError("Encrypt operation failed", err, t)
-	oid := encryptResponse.ObjectId
+	oid := encryptResponse.ObjectID
 
 	// Try to add permissions for a non-existing user
-	_, err = client.AddPermission(oid, nonExistingUser)
+	err = client.AddPermission(oid, nonExistingUser)
 	failOnSuccess("Shouldn't able to add user that does not exist!", err, t)
 }
