@@ -18,25 +18,29 @@
 package grpce2e
 
 import (
-	"reflect"
 	"testing"
+
+	"context"
+	"reflect"
+
+	coreclient "github.com/cyber-crypt-com/encryptonize-core/client"
 )
 
 // Test that a user can remove themselves from the object ACL and cannot access the object afterwards
 func TestRetrieveSameUserWithoutPermissions(t *testing.T) {
-	client, err := NewClient(endpoint, certPath)
+	client, err := coreclient.NewClient(context.Background(), endpoint, certPath)
 	failOnError("Could not create client", err, t)
-	defer closeClient(client, t)
+	defer client.Close()
 
-	_, err = client.LoginUser(uid, pwd)
+	err = client.LoginUser(uid, pwd)
 	failOnError("Could not log in user", err, t)
 
 	storeResponse, err := client.Store([]byte("foo"), []byte("bar"))
 	failOnError("Store operation failed", err, t)
-	oid := storeResponse.ObjectId
+	oid := storeResponse.ObjectID
 
 	// Remove self from object ACL
-	_, err = client.RemovePermission(oid, uid)
+	err = client.RemovePermission(oid, uid)
 	failOnError("Removing permissions from self failed", err, t)
 
 	// Try to fetch object without permissions
@@ -47,17 +51,17 @@ func TestRetrieveSameUserWithoutPermissions(t *testing.T) {
 // Test that a stored object can be retrieved by another user with permissions
 func TestShareObjectWithUser(t *testing.T) {
 	// Create another user to share the object with
-	client, err := NewClient(endpoint, certPath)
+	client, err := coreclient.NewClient(context.Background(), endpoint, certPath)
 	failOnError("Could not create client", err, t)
-	defer closeClient(client, t)
+	defer client.Close()
 
-	_, err = client.LoginUser(uid, pwd)
+	err = client.LoginUser(uid, pwd)
 	failOnError("Could not log in user", err, t)
 
 	createUserResponse, err := client.CreateUser(protoUserScopes)
 	failOnError("Create user request failed", err, t)
 
-	uid2 := createUserResponse.UserId
+	uid2 := createUserResponse.UserID
 	pwd2 := createUserResponse.Password
 
 	// Store an object
@@ -65,21 +69,21 @@ func TestShareObjectWithUser(t *testing.T) {
 	associatedData := []byte("bar")
 	storeResponse, err := client.Store(plaintext, associatedData)
 	failOnError("Store operation failed", err, t)
-	oid := storeResponse.ObjectId
+	oid := storeResponse.ObjectID
 
 	// Add another user to object permission list
-	_, err = client.AddPermission(oid, uid2)
+	err = client.AddPermission(oid, uid2)
 	failOnError("Add permission request failed", err, t)
 
 	// Try to retrieve object with another user
-	_, err = client.LoginUser(uid2, pwd2)
+	err = client.LoginUser(uid2, pwd2)
 	failOnError("Could not log in user", err, t)
 
 	_, err = client.Retrieve(oid)
 	failOnError("Authorized user could not fetch object created by another user", err, t)
 
 	// Try to retrieve object with the original user
-	_, err = client.LoginUser(uid, pwd)
+	err = client.LoginUser(uid, pwd)
 	failOnError("Could not log in user", err, t)
 
 	_, err = client.Retrieve(oid)
@@ -89,17 +93,17 @@ func TestShareObjectWithUser(t *testing.T) {
 // Test that a stored object cannot be retrieved by another user without permissions
 func TestRetrieveWithoutPermissions(t *testing.T) {
 	// Create another user to share the object with
-	client, err := NewClient(endpoint, certPath)
+	client, err := coreclient.NewClient(context.Background(), endpoint, certPath)
 	failOnError("Could not create client", err, t)
-	defer closeClient(client, t)
+	defer client.Close()
 
-	_, err = client.LoginUser(uid, pwd)
+	err = client.LoginUser(uid, pwd)
 	failOnError("Could not log in user", err, t)
 
 	createUserResponse, err := client.CreateUser(protoUserScopes)
 	failOnError("Create user request failed", err, t)
 
-	uid2 := createUserResponse.UserId
+	uid2 := createUserResponse.UserID
 	pwd2 := createUserResponse.Password
 
 	// Store an object
@@ -107,10 +111,10 @@ func TestRetrieveWithoutPermissions(t *testing.T) {
 	associatedData := []byte("bar")
 	storeResponse, err := client.Store(plaintext, associatedData)
 	failOnError("Store operation failed", err, t)
-	oid := storeResponse.ObjectId
+	oid := storeResponse.ObjectID
 
 	// Try to retrieve object with another user
-	_, err = client.LoginUser(uid2, pwd2)
+	err = client.LoginUser(uid2, pwd2)
 	failOnError("Could not log in user", err, t)
 
 	_, err = client.Retrieve(oid)
@@ -121,24 +125,24 @@ func TestRetrieveWithoutPermissions(t *testing.T) {
 // If user A grants access to user B, then user B should be able to grant access to user C
 func TestPermissionTransitivity(t *testing.T) {
 	// Create admin client for user creation
-	client, err := NewClient(endpoint, certPath)
+	client, err := coreclient.NewClient(context.Background(), endpoint, certPath)
 	failOnError("Could not create client", err, t)
-	defer closeClient(client, t)
+	defer client.Close()
 
-	_, err = client.LoginUser(uid, pwd)
+	err = client.LoginUser(uid, pwd)
 	failOnError("Could not log in user", err, t)
 
 	// Create users B and C
 	createUserResponse, err := client.CreateUser(protoUserScopes)
 	failOnError("Create user request failed", err, t)
 
-	uid2 := createUserResponse.UserId
+	uid2 := createUserResponse.UserID
 	pwd2 := createUserResponse.Password
 
 	createUserResponse, err = client.CreateUser(protoUserScopes)
 	failOnError("Create user request failed", err, t)
 
-	uid3 := createUserResponse.UserId
+	uid3 := createUserResponse.UserID
 	pwd3 := createUserResponse.Password
 
 	// Store an object
@@ -146,21 +150,21 @@ func TestPermissionTransitivity(t *testing.T) {
 	associatedData := []byte("bar")
 	storeResponse, err := client.Store(plaintext, associatedData)
 	failOnError("Store operation failed", err, t)
-	oid := storeResponse.ObjectId
+	oid := storeResponse.ObjectID
 
 	// Grant permissions to user B
-	_, err = client.AddPermission(oid, uid2)
+	err = client.AddPermission(oid, uid2)
 	failOnError("Add permission request failed", err, t)
 
 	// Use user B to grant permissions to user C
-	_, err = client.LoginUser(uid2, pwd2)
+	err = client.LoginUser(uid2, pwd2)
 	failOnError("Could not log in user", err, t)
 
-	_, err = client.AddPermission(oid, uid3)
+	err = client.AddPermission(oid, uid3)
 	failOnError("Add permission request failed", err, t)
 
 	// Check that user C has access to object
-	_, err = client.LoginUser(uid3, pwd3)
+	err = client.LoginUser(uid3, pwd3)
 	failOnError("Could not log in user", err, t)
 
 	_, err = client.Retrieve(oid)
@@ -171,18 +175,18 @@ func TestPermissionTransitivity(t *testing.T) {
 // and that add/remove permission inflicts the outcome of get permissions
 func TestGetPermissions(t *testing.T) {
 	// Create admin client for user creation
-	client, err := NewClient(endpoint, certPath)
+	client, err := coreclient.NewClient(context.Background(), endpoint, certPath)
 	failOnError("Could not create client", err, t)
-	defer closeClient(client, t)
+	defer client.Close()
 
-	_, err = client.LoginUser(uid, pwd)
+	err = client.LoginUser(uid, pwd)
 	failOnError("Could not log in user", err, t)
 
 	// Create user 2
 	createUserResponse, err := client.CreateUser(protoUserScopes)
 	failOnError("Create user request failed", err, t)
 
-	uid2 := createUserResponse.UserId
+	uid2 := createUserResponse.UserID
 	pwd2 := createUserResponse.Password
 
 	// Store an object
@@ -190,66 +194,66 @@ func TestGetPermissions(t *testing.T) {
 	associatedData := []byte("bar")
 	storeResponse, err := client.Store(plaintext, associatedData)
 	failOnError("Store operation failed", err, t)
-	oid := storeResponse.ObjectId
+	oid := storeResponse.ObjectID
 
 	//Check that user 2 cannot get permissions from object, without permissions
-	_, err = client.LoginUser(uid2, pwd2)
+	err = client.LoginUser(uid2, pwd2)
 	failOnError("Could not log in user", err, t)
 
 	_, err = client.GetPermissions(oid)
 	failOnSuccess("Unauthorized user should not be able to access object permissions", err, t)
 
 	// Grant permissions to user 2
-	_, err = client.LoginUser(uid, pwd)
+	err = client.LoginUser(uid, pwd)
 	failOnError("Could not log in user", err, t)
 
-	_, err = client.AddPermission(oid, uid2)
+	err = client.AddPermission(oid, uid2)
 	failOnError("Add permission request failed", err, t)
 
 	getPermissionsResponse1, err := client.GetPermissions(oid)
 	failOnError("Could not get permissions", err, t)
 
-	_, err = client.LoginUser(uid2, pwd2)
+	err = client.LoginUser(uid2, pwd2)
 	failOnError("Could not log in user", err, t)
 
 	getPermissionsResponse2, err := client.GetPermissions(oid)
 	failOnError("Could not get permissions", err, t)
 
 	// Check that permissions response contains the right uids
-	ok := find(getPermissionsResponse1.GroupIds, uid)
+	ok := find(getPermissionsResponse1.GroupIDs, uid)
 	if !ok {
-		t.Fatalf("Couldn't find %v in %v", uid, getPermissionsResponse1.GroupIds)
+		t.Fatalf("Couldn't find %v in %v", uid, getPermissionsResponse1.GroupIDs)
 	}
-	ok = find(getPermissionsResponse1.GroupIds, uid2)
+	ok = find(getPermissionsResponse1.GroupIDs, uid2)
 	if !ok {
-		t.Fatalf("Couldn't find %v in %v", uid, getPermissionsResponse1.GroupIds)
+		t.Fatalf("Couldn't find %v in %v", uid, getPermissionsResponse1.GroupIDs)
 	}
 
-	if !reflect.DeepEqual(getPermissionsResponse1.GroupIds, getPermissionsResponse2.GroupIds) {
-		t.Fatalf("Permissions aren't the same: %v vs %v", getPermissionsResponse1.GroupIds, getPermissionsResponse2.GroupIds)
+	if !reflect.DeepEqual(getPermissionsResponse1.GroupIDs, getPermissionsResponse2.GroupIDs) {
+		t.Fatalf("Permissions aren't the same: %v vs %v", getPermissionsResponse1.GroupIDs, getPermissionsResponse2.GroupIDs)
 	}
 
 	// Remove user 2
-	_, err = client.LoginUser(uid, pwd)
+	err = client.LoginUser(uid, pwd)
 	failOnError("Could not log in user", err, t)
 
-	_, err = client.RemovePermission(oid, uid2)
+	err = client.RemovePermission(oid, uid2)
 	failOnError("Could not remove permissions", err, t)
 
 	// Check that permissions have been removed
 	getPermissionsResponse1, err = client.GetPermissions(oid)
 	failOnError("Could not get permissions", err, t)
-	ok = find(getPermissionsResponse1.GroupIds, uid)
+	ok = find(getPermissionsResponse1.GroupIDs, uid)
 	if !ok {
-		t.Fatalf("Couldn't find %v in %v", uid, getPermissionsResponse1.GroupIds)
+		t.Fatalf("Couldn't find %v in %v", uid, getPermissionsResponse1.GroupIDs)
 	}
-	ok = find(getPermissionsResponse1.GroupIds, uid2)
+	ok = find(getPermissionsResponse1.GroupIDs, uid2)
 	if ok {
-		t.Fatalf("Found %v in %v", uid, getPermissionsResponse1.GroupIds)
+		t.Fatalf("Found %v in %v", uid, getPermissionsResponse1.GroupIDs)
 	}
 
 	// Check that user 2 doesn't have permissions
-	_, err = client.LoginUser(uid2, pwd2)
+	err = client.LoginUser(uid2, pwd2)
 	failOnError("Could not log in user", err, t)
 
 	_, err = client.Retrieve(oid)
@@ -260,18 +264,18 @@ func TestGetPermissions(t *testing.T) {
 func TestAddPermissionNoTargetUser(t *testing.T) {
 	nonExistingUser := "00000000-0000-0000-0000-000000000000"
 
-	client, err := NewClient(endpoint, certPath)
+	client, err := coreclient.NewClient(context.Background(), endpoint, certPath)
 	failOnError("Could not create client", err, t)
-	defer closeClient(client, t)
+	defer client.Close()
 
-	_, err = client.LoginUser(uid, pwd)
+	err = client.LoginUser(uid, pwd)
 	failOnError("Could not log in user", err, t)
 
 	storeResponse, err := client.Store([]byte("foo"), []byte("bar"))
 	failOnError("Store operation failed", err, t)
-	oid := storeResponse.ObjectId
+	oid := storeResponse.ObjectID
 
 	// Try to add permissions for a non-existing user
-	_, err = client.AddPermission(oid, nonExistingUser)
+	err = client.AddPermission(oid, nonExistingUser)
 	failOnSuccess("Shouldn't able to add user that does not exist!", err, t)
 }
