@@ -18,22 +18,33 @@
 package grpce2e
 
 import (
-	"bytes"
-	"fmt"
-	"sync"
 	"testing"
 	"testing/quick"
 
-	"encryption-service/impl/crypt"
+	"bytes"
+	"context"
+	"crypto/rand"
+	"fmt"
+	"sync"
+
+	coreclient "github.com/cyber-crypt-com/encryptonize-core/client"
 )
+
+func Random(n int) ([]byte, error) {
+	bytes := make([]byte, n)
+	if _, err := rand.Read(bytes); err != nil {
+		return nil, err
+	}
+	return bytes, nil
+}
 
 // Test the we can store an object and retrieve it later
 func TestStoreAndRetrieve(t *testing.T) {
-	client, err := NewClient(endpoint, certPath)
+	client, err := coreclient.NewClient(context.Background(), endpoint, certPath)
 	failOnError("Could not create client", err, t)
-	defer closeClient(client, t)
+	defer client.Close()
 
-	_, err = client.LoginUser(uid, pwd)
+	err = client.LoginUser(uid, pwd)
 	failOnError("Could not log in user", err, t)
 
 	plaintext := []byte("foo")
@@ -41,7 +52,7 @@ func TestStoreAndRetrieve(t *testing.T) {
 
 	storeResponse, err := client.Store(plaintext, associatedData)
 	failOnError("Store operation failed", err, t)
-	oid := storeResponse.ObjectId
+	oid := storeResponse.ObjectID
 	retrieveResponse, err := client.Retrieve(oid)
 	failOnError("Retrieve operation failed", err, t)
 
@@ -56,11 +67,11 @@ func TestStoreAndRetrieve(t *testing.T) {
 
 // Test that we can store the same object/data multiple times and still retrieve it
 func TestStoreSameObjectMultipleTimes(t *testing.T) {
-	client, err := NewClient(endpoint, certPath)
+	client, err := coreclient.NewClient(context.Background(), endpoint, certPath)
 	failOnError("Could not create client", err, t)
-	defer closeClient(client, t)
+	defer client.Close()
 
-	_, err = client.LoginUser(uid, pwd)
+	err = client.LoginUser(uid, pwd)
 	failOnError("Could not log in user", err, t)
 
 	plaintext := []byte("foo")
@@ -69,7 +80,7 @@ func TestStoreSameObjectMultipleTimes(t *testing.T) {
 	for i := 0; i < 5; i++ {
 		storeResponse, err := client.Store(plaintext, associatedData)
 		failOnError("Store operation failed", err, t)
-		oid := storeResponse.ObjectId
+		oid := storeResponse.ObjectID
 		retrieveResponse, err := client.Retrieve(oid)
 		failOnError("Retrieve operation failed", err, t)
 
@@ -85,11 +96,11 @@ func TestStoreSameObjectMultipleTimes(t *testing.T) {
 
 // Test retrieving object with invalid oid
 func TestRetrieveBadOid(t *testing.T) {
-	client, err := NewClient(endpoint, certPath)
+	client, err := coreclient.NewClient(context.Background(), endpoint, certPath)
 	failOnError("Could not create client", err, t)
-	defer closeClient(client, t)
+	defer client.Close()
 
-	_, err = client.LoginUser(uid, pwd)
+	err = client.LoginUser(uid, pwd)
 	failOnError("Could not log in user", err, t)
 
 	oid := "bad"
@@ -99,11 +110,11 @@ func TestRetrieveBadOid(t *testing.T) {
 
 // Test multiple retrieve operations of the same object
 func TestMultipleStoreRetrieve(t *testing.T) {
-	client, err := NewClient(endpoint, certPath)
+	client, err := coreclient.NewClient(context.Background(), endpoint, certPath)
 	failOnError("Could not create client", err, t)
-	defer closeClient(client, t)
+	defer client.Close()
 
-	_, err = client.LoginUser(uid, pwd)
+	err = client.LoginUser(uid, pwd)
 	failOnError("Could not log in user", err, t)
 
 	plaintext := []byte("foo")
@@ -111,7 +122,7 @@ func TestMultipleStoreRetrieve(t *testing.T) {
 
 	storeResponse, err := client.Store(plaintext, associatedData)
 	failOnError("Store operation failed", err, t)
-	oid := storeResponse.ObjectId
+	oid := storeResponse.ObjectID
 
 	// Retrieve the object multiple times
 	for i := 0; i < 10; i++ {
@@ -130,11 +141,11 @@ func TestMultipleStoreRetrieve(t *testing.T) {
 
 // Test that an older object can be retrieved after storing multiple objects.
 func TestRetrieveOlderObject(t *testing.T) {
-	client, err := NewClient(endpoint, certPath)
-	defer closeClient(client, t)
+	client, err := coreclient.NewClient(context.Background(), endpoint, certPath)
 	failOnError("Could not create client", err, t)
+	defer client.Close()
 
-	_, err = client.LoginUser(uid, pwd)
+	err = client.LoginUser(uid, pwd)
 	failOnError("Could not log in user", err, t)
 
 	// Store first object
@@ -142,20 +153,20 @@ func TestRetrieveOlderObject(t *testing.T) {
 	associatedData := []byte("bar1")
 	storeResponse, err := client.Store(plaintext, associatedData)
 	failOnError("Could not create first object", err, t)
-	oid := storeResponse.ObjectId
+	oid := storeResponse.ObjectID
 
 	// Store some other objects
 	plaintext2 := []byte("foo2")
 	associatedData2 := []byte("bar2")
 	storeResponse2, err := client.Store(plaintext2, associatedData2)
 	failOnError("Could not create second object", err, t)
-	_, err = client.Retrieve(storeResponse2.ObjectId)
+	_, err = client.Retrieve(storeResponse2.ObjectID)
 	failOnError("Could not retrieve object", err, t)
 	plaintext3 := []byte("foo3")
 	associatedData3 := []byte("bar3")
 	storeResponse3, err := client.Store(plaintext3, associatedData3)
 	failOnError("Could not create third object", err, t)
-	_, err = client.Retrieve(storeResponse3.ObjectId)
+	_, err = client.Retrieve(storeResponse3.ObjectID)
 	failOnError("Could not retrieve object", err, t)
 
 	// Retrieve first object
@@ -173,42 +184,42 @@ func TestRetrieveOlderObject(t *testing.T) {
 
 // Test that a user can store and retrieve a bigger object
 func TestStoreRetrieveBigObject(t *testing.T) {
-	client, err := NewClient(endpoint, certPath)
+	client, err := coreclient.NewClient(context.Background(), endpoint, certPath)
 	failOnError("Could not create client", err, t)
-	defer closeClient(client, t)
+	defer client.Close()
 
-	_, err = client.LoginUser(uid, pwd)
+	err = client.LoginUser(uid, pwd)
 	failOnError("Could not log in user", err, t)
 
-	object, err := crypt.Random(1024 * 1024 * 2) // 2 MB object
+	object, err := Random(1024 * 1024 * 2) // 2 MB object
 	failOnError("Could not create random object", err, t)
-	associatedData, err := crypt.Random(1024 * 1024 * 1) // 1 MB AD
+	associatedData, err := Random(1024 * 1024 * 1) // 1 MB AD
 	failOnError("Could not create random object", err, t)
 
 	storeResponse, err := client.Store(object, associatedData)
 	failOnError("Store operation failed", err, t)
-	oid := storeResponse.ObjectId
+	oid := storeResponse.ObjectID
 	_, err = client.Retrieve(oid)
 	failOnError("Retrieve operation failed", err, t)
 }
 
 // Tests random sizes to store and retrieve
 func TestStoreRetrieveRandomSizes(t *testing.T) {
-	client, err := NewClient(endpoint, certPath)
+	client, err := coreclient.NewClient(context.Background(), endpoint, certPath)
 	failOnError("Could not create client", err, t)
-	defer closeClient(client, t)
+	defer client.Close()
 
-	_, err = client.LoginUser(uid, pwd)
+	err = client.LoginUser(uid, pwd)
 	failOnError("Could not log in user", err, t)
 
 	storeRetrieve := func(obsz uint32, adsz uint32) bool {
 		// if the modulo is changed here, it also has to be changed in the error reporting below
-		object, err := crypt.Random(int(obsz % 4096)) // at most 4kB object
+		object, err := Random(int(obsz % 4096)) // at most 4kB object
 		if err != nil {
 			return false
 		}
 
-		associatedData, err := crypt.Random(int(adsz % 4096)) // at most 4kB AD
+		associatedData, err := Random(int(adsz % 4096)) // at most 4kB AD
 		if err != nil {
 			return false
 		}
@@ -218,7 +229,7 @@ func TestStoreRetrieveRandomSizes(t *testing.T) {
 			return false
 		}
 
-		oid := storeResponse.ObjectId
+		oid := storeResponse.ObjectID
 		_, err = client.Retrieve(oid)
 
 		return err == nil
@@ -265,24 +276,24 @@ func TestConcurrentStoreRetrieve(t *testing.T) {
 		wg.Add(1)
 		go func(wg *sync.WaitGroup) {
 			defer wg.Done()
-			plaintext, err := crypt.Random(1024) // 1KB
-			if err != nil {                      // Bail out early on error
+			plaintext, err := Random(1024) // 1KB
+			if err != nil {                // Bail out early on error
 				globErr.append(fmt.Errorf("Couldn't create object: %v", err))
 				return
 			}
-			associatedData, err := crypt.Random(1024) // 1KB
-			if err != nil {                           // Bail out early on error
+			associatedData, err := Random(1024) // 1KB
+			if err != nil {                     // Bail out early on error
 				globErr.append(fmt.Errorf("Couldn't create AD: %v", err))
 				return
 			}
-			client, err := NewClient(endpoint, certPath)
-			defer closeClient(client, t)
+			client, err := coreclient.NewClient(context.Background(), endpoint, certPath)
 			if err != nil { // Bail out early on error
 				globErr.append(fmt.Errorf("Couldn't create client: %v", err))
 				return
 			}
+			defer client.Close()
 
-			_, err = client.LoginUser(uid, pwd)
+			err = client.LoginUser(uid, pwd)
 			if err != nil { // Bail out early on error
 				globErr.append(fmt.Errorf("Couldn't log in user: %v", err))
 				return
@@ -293,7 +304,7 @@ func TestConcurrentStoreRetrieve(t *testing.T) {
 				globErr.append(fmt.Errorf("Couldn't create store object: %v", err))
 				return
 			}
-			oid := storeResponse.ObjectId
+			oid := storeResponse.ObjectID
 			retrieveResponse, err := client.Retrieve(oid)
 			if err != nil { // Bail out early on error
 				globErr.append(fmt.Errorf("Couldn't retrieve object: %v", err))
@@ -314,11 +325,11 @@ func TestConcurrentStoreRetrieve(t *testing.T) {
 
 // Test storage and retrieval of objects with no data and no associated data
 func TestStoreRetrieveEmptyObjects(t *testing.T) {
-	client, err := NewClient(endpoint, certPath)
+	client, err := coreclient.NewClient(context.Background(), endpoint, certPath)
 	failOnError("Could not create client", err, t)
-	defer closeClient(client, t)
+	defer client.Close()
 
-	_, err = client.LoginUser(uid, pwd)
+	err = client.LoginUser(uid, pwd)
 	failOnError("Could not log in user", err, t)
 
 	emptyPlaintext := []byte("")
@@ -329,7 +340,7 @@ func TestStoreRetrieveEmptyObjects(t *testing.T) {
 	// Store empty object and associated data
 	storeResponse, err := client.Store(emptyPlaintext, emptyAssociatedData)
 	failOnError("Stroing of empty object and associated data failed", err, t)
-	oid := storeResponse.ObjectId
+	oid := storeResponse.ObjectID
 	retrieveResponse, err := client.Retrieve(oid)
 	failOnError("Retrieve empty object failed", err, t)
 
@@ -344,7 +355,7 @@ func TestStoreRetrieveEmptyObjects(t *testing.T) {
 	// Store non-empty object and empty associated data
 	storeResponse, err = client.Store(nonEmptyPlaintext, emptyAssociatedData)
 	failOnError("Stroing of non-empty object and empty associated data failed", err, t)
-	oid = storeResponse.ObjectId
+	oid = storeResponse.ObjectID
 	retrieveResponse, err = client.Retrieve(oid)
 	failOnError("Retrieve object failed", err, t)
 
@@ -359,7 +370,7 @@ func TestStoreRetrieveEmptyObjects(t *testing.T) {
 	// Store empty object and non-empty associated data
 	storeResponse, err = client.Store(emptyPlaintext, nonEmptyAssociatedData)
 	failOnError("Stroing of empty object and associated data failed", err, t)
-	oid = storeResponse.ObjectId
+	oid = storeResponse.ObjectID
 	retrieveResponse, err = client.Retrieve(oid)
 	failOnError("Retrieve empty object failed", err, t)
 
